requirements.txt
Flask
nltk
newspaper3k